+++ bin/src/api/user/user.controller.ts
@@ -5,6 +5,7 @@ import { createDirectory, getUser, updateUser } from './user.service';
 export const UserRouter = express.Router();
 
 UserRouter.use(loginValidation);
-UserRouter.get('/me', getUser);
-UserRouter.patch('/me', editUserValidation, updateUser);
-UserRouter.post('/me', createDirectory);
+UserRouter.route('/me')
+  .get(getUser)
+  .patch(editUserValidation, updateUser)
+  .post(createDirectory);