+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Patch 1.0.2
+### Added: 
+* Option to delete server logs
+
+### Updated:
+* Improved logging
+
+### Fixed:
+* After inserting command, procces had not been exiting
+
+`03-06-2024`
+
 ## Patch 1.0.1
 ### Added:
 * Now user can delete his storage, and recreate it again